pyproject.toml
[project]
name = "pycreate2"
version = "0.9"
description = "A library to control iRobot Create 2 with python"
readme = "readme.md"
homepage = "https://pypi.org/project/pycreate2/"
repository = "https://github.com/GDPB3/pycreate2"
keywords = ["roomba","irobot", "create 2", "api", "framework", "library", "robotics", "robot"]
license = { file = "LICENSE" }
authors = [
    { name = "walchko", email = "walchko@users.noreply.github.com" },
    { name = "luengor", email = "aluengor0001@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
    "pytest>=8.4.2",
]

[project.scripts]
create_monitor = "pycreate2.scripts.create_monitor:main"
create_reset = "pycreate2.scripts.create_reset:main"
create_shutdown = "pycreate2.scripts.create_shutdown:main"

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"
